refactor import in init file. organize zaptoken
